+++ src/main/java/subway/application/feepolicy/DefaultFeeCalculator.java
@@ -3,15 +3,18 @@ package subway.application.feepolicy;
 import org.springframework.stereotype.Component;
 
 @Component
-public class DefaultFeePolicy implements FeePolicy {
+public class DefaultFeeCalculator implements FeeCalculator {
+    private static final int DEFAULT_FEE = 1250;
     private static final int BASIC_DISTANCE = 10;
     private static final int ADDITIONAL_DISTANCE = 50;
     private static final int ADDITIONAL_FEE = 100;
+    private static final double UNIT_DISTANCE = 5d;
+    private static final double LONGER_UNIT_DISTANCE = 8d;
 
     @Override
     public int calculateFee(final int distance) {
 
-        int fee = 1250;
+        int fee = DEFAULT_FEE;
         int distanceBeforeFifty = 0;
         int distanceAfterFifty = 0;
 
@@ -22,10 +25,10 @@ public class DefaultFeePolicy implements FeePolicy {
         }
 
         if (distanceBeforeFifty > 0) {
-            fee += Math.ceil(distanceBeforeFifty / 5d) * ADDITIONAL_FEE;
+            fee += Math.ceil(distanceBeforeFifty / UNIT_DISTANCE) * ADDITIONAL_FEE;
         }
         if (distanceAfterFifty > 0) {
-            fee += Math.ceil(distanceAfterFifty / 8d) * ADDITIONAL_FEE;
+            fee += Math.ceil(distanceAfterFifty / LONGER_UNIT_DISTANCE) * ADDITIONAL_FEE;
         }
 
         return fee;
