+++ src/test/java/subway/application/DefaultFeePolicyTest.java
@@ -3,7 +3,7 @@ package subway.application;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import subway.application.feepolicy.DefaultFeePolicy;
+import subway.application.feepolicy.DefaultFeeCalculator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -13,7 +13,7 @@ class DefaultFeePolicyTest {
     @ParameterizedTest
     @CsvSource(value = {"9, 1250", "10, 1250", "14, 1350", "15, 1350", "58, 2150"})
     void calculateFee(int distance, int fee) {
-        final DefaultFeePolicy policy = new DefaultFeePolicy();
+        final DefaultFeeCalculator policy = new DefaultFeeCalculator();
 
         final int result = policy.calculateFee(distance);
